requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
opengradient==0.7.4
web3==7.14.1
gunicorn==23.0.0
